try moving tensor product in pec
